requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pyyaml==6.0.1
python-jose==3.3.0
python-dotenv==1.0.0
httpx==0.25.1
PyGithub==2.6.0
pydantic==2.4.2 